+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Rename `DefaultStatsCollector` to `StatsCollector`.
+
 ### Fixed
 
 ## [3.0.0-beta.1] - 2022-02-23